Move SheetsService exceptions to own folder
