+++ DiscordBot.DataAccess/Exceptions/EventInitialisationException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DiscordBot.DataAccess
+namespace DiscordBot.DataAccess.Exceptions
 {
     class EventInitialisationException : Exception
     {
