+++ DiscordBot.DataAccess/Exceptions/EventNotFoundException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DiscordBot.DataAccess
+namespace DiscordBot.DataAccess.Exceptions
 {
     public class EventNotFoundException : Exception
     {
